Add changes syncing to the backend from indexeddb.
